+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Latest
  * Added a few tests for the sunpy.roi module
  * Refactored mapcube co-alignment functionality.
  * Removed sample data from distribution and added ability to download sample files
+ * Added passing of astropy.units.Unit as default_waveunit to sunpy.database.Database
 
 0.5.0
 -----